move infowindow manager to deep insights integration and its own file
