+++ lib/assets/javascripts/cartodb3/deep-insights-integrations.js
@@ -1,3 +1,5 @@
+var InfowindowManager = require('./infowindow-manager');
+
 /**
  * Integration between various data collections/models with cartodb.js and deep-insights.
  */
@@ -29,6 +31,13 @@ var F = function (opts) {
   opts.widgetDefinitionsCollection.on('add remove reset', this._invalidateSize, this);
 
   this._analysisDefinitionsCollection.each(this._analyseDefinition, this);
+
+  var infowindowManager = new InfowindowManager({
+    visMap: this.visMap(),
+    layerDefinitionsCollection: opts.layerDefinitionsCollection
+  });
+
+  infowindowManager.initBinds();
 };
 
 F.prototype._onAnalysisDefinitionNodeChanged = function () {
