+++ lib/assets/javascripts/cartodb3/editor.js
@@ -41,52 +41,6 @@ var visualizationData = window.visualizationData;
 var layersData = window.layersData;
 var analysesData = window.analysesData;
 
-function InfowindowLink (layerDef, visMap) {
-  if (!layerDef.infowindowModel) return;
-
-  function onChange (attrs) {
-    visMap.getLayerById(layerDef.id).infowindow.update(attrs.toJSON());
-  }
-
-  layerDef.infowindowModel.bind('change', onChange);
-
-  layerDef.bind('destroy', function () {
-    layerDef.infowindowModel.unbind('change', onChange);
-  });
-}
-
-function TooltipLink (layerDef, visMap) {
-  if (!layerDef.tooltipModel) return;
-
-  function onChange (attrs) {
-    visMap.getLayerById(layerDef.id).tooltip.update(attrs.toJSON());
-  }
-
-  layerDef.tooltipModel.bind('change', onChange);
-
-  layerDef.bind('destroy', function () {
-    layerDef.tooltipModel.unbind('change', onChange);
-  });
-}
-
-function InfowindowManager (opts) {
-  opts.layerDefinitionsCollection.on('add', function (layerDef) {
-    InfowindowLink(layerDef, opts.vis.map);
-    TooltipLink(layerDef, opts.vis.map);
-  });
-
-  function update () {
-    opts.layerDefinitionsCollection.each(function (layerDef) {
-      InfowindowLink(layerDef, opts.vis.map);
-      TooltipLink(layerDef, opts.vis.map);
-    });
-  }
-
-  this.initBinds = update.bind(this);
-
-  opts.layerDefinitionsCollection.on('reset', update);
-}
-
 cdb.god = new cdb.core.Model(); // To be replaced
 
 var configModel = new ConfigModel(
@@ -151,13 +105,7 @@ deepInsights.createDashboard('#dashboard', vizJSON, {
 
   dashboardView.add_related_model(editorModel);
 
-  var infowindowManager = new InfowindowManager({
-    vis: vis,
-    layerDefinitionsCollection: layerDefinitionsCollection
-  });
-
   layerDefinitionsCollection.resetByLayersData(layersData);
-  infowindowManager.initBinds();
 
   var widgetDefinitionsCollection = new WidgetDefinitionsCollection(null, {
     configModel: configModel,
